Create `SBase` validation module.
